Arreglo problemas con fifo tanto para el Cocinero como para lista en el living, ahora esta andando correctamente

+++ src/com/fiuba/resto/process/action/AddDinerToLivingLineAction.h
@@ -26,7 +26,7 @@ private:
 public:
 	AddDinerToLivingLineAction();
 	virtual ~AddDinerToLivingLineAction();
-	void run(unsigned int dinerPid);
+	void run(unsigned long dinerPid);
 };
 
 #endif /* ADD_DINER_TO_LIVING_ACTION_H_ */
